This check in should complete server side for #10 #12 #18 - Updated user db for lat long - Added something to calculate user distances - Added rest api to be able to get list of users near a location - Updated battle & user related tests

+++ server/database/migrations/2014_10_12_000000_create_users_table.php
@@ -18,8 +18,8 @@ class CreateUsersTable extends Migration
             $table->string('name');
             $table->string('email')->unique();
             $table->string('password');
-            $table->float('latitude')->default(0);
-            $table->float('longitude')->default(0);
+            $table->decimal('latitude',10,7)->default(0);
+            $table->decimal('longitude',10,7)->default(0);
             $table->integer('battles_won')->unsigned()->default(0);
             $table->rememberToken();
             $table->timestamps();
